Right now, I've been taking a break by reading up on the DenseNet paper.
https://arxiv.org/abs/1608.06993
Densely Connected Convolutional Networks

This paper is actually a great motivator for me to implement my own generic gemm and convolution because those kernels would be a tailor made use for them. And given how good these architectures are, I should definitely consider them past 4 layers.

But it is not a priority.

What I need right now is to master just a single recurrent layer.

I did well by starting with mi+ln+relu in this respect.

https://arxiv.org/abs/1707.06990
Memory-Efficient Implementation of DenseNets

Let me read this and then I will work a little on the tutorial. It won't be long now until I start doing things that are really important.