Here it seems a good spot to take a little break after all the effort of the past few days.
After that, I'll start a new project and slowly get to work with poker and tabular RL. I already did this before in F#, and I quite like the experience. Unfortunately, I do not have much choice, but to do it from scratch.

This will take me a while, but as usual, I'll aim for steady improvements each and every day.

I wonder what I will get when I finally plug a RNN in?