Actually this is pretty complicated. Yeah, no matter what, making a game is complicated. It was hard last time and I am not exactly sure what the architecture of the library should be.
I know how I did it last time, and I have an extra list of requirement this time - I want the library to be flexible enough to support any kind of card game.

But now I am wondering what the actual design should be.

Unlike last time, I think I should not give the bots any state. Instead I should do a global controller that takes in messages. That would be the best thing.

Let me try some things. I am starting to feel the inspiration coming on. Something like this would be hard for me before, but a task like this is only of moderate difficulty to me now.

I won't be spending forever on it.