I've thought of what I want to do from here on out. GUIs. The poker game won't be a problem by itself. The feedback I've received yesterday was extremely favorable. Building up the RL aspects will be pretty much trivial.
But the damn GUIs on the other hand. Yeah, they will be a problem. I did a lot of work on this in 2016, but unfortunately, I did not master the subject, so that will have to be done now. I think I will ask on the F# sub.

Also to start the day off, I do feel like going through the programming for correctness course.

I guess I can take it easy for a while, to do some studying again.