Right now I am really doing some soul searching. Looking at my old code I really feel bad about abandoning it.
But I do not think I want GUI dependencies right there in the game any longer. It was hard to do in F# last time, it will be 5x hard in Spiral without IDE support and macros for interop.

The real reason why I wanted it last time was to help debugging and give me experience some sophistication, but GUIs are such a bother.

If I want a tool to review old games, I could just build a tool to visualize it in F#. The communication would be done via text, but I am good at making parsers at this point. And this would be the ideal approach from a certain angle. The RL and DL stuff would be completely isolated from the rest - the GUI code would create messy coupling, and the GUI stuff would have its own dedicated machinery.

Related to this, there is also a separate concern. If I am having difficulty making a simple GUI for this, what will I do in the later stages when it comes to doing hacker-tier stuff like granting a trained RNN access to the entire system? At least, I'll need to do things like capture the screen and feed it into it and respond to its outputs.

But looking at it reasonably, I do not really want GUIs everywhere as they are so heavyweight. I certainly do not want concurrent code right smack dab in the middle of the game engine. I'd be perfectly content if I could somehow get the different processes to communicate properly. They do not have to be processes though, that might be too much. I might be able to get away with actors and message passing.

So maybe even if I decide to skip on doing GUIs in Spiral, I could maybe look into how to make actor concurrency and message passing work?

So how is this for a goal - I'll make a poker game in Spiral and a visualizer in F# as separate applications. This would have synergies in that if I put a GUI in the game it would only be useful for that, but a visualizer would be good for not just my own games, but even online poker sites. They all allow saving hand histories so I will be making parsers for them at any rate.

After making both is done, I'll somehow find a way to merge them using actors and message passing rather than having them communicate via disk. I think this seems like a doable plan. I mean the GUI being a separate actor from the game should be viable somehow.

So with that in mind how about I first finish the game and the tabular agents and then get to work on the rest? My instinct as a programmer is telling me that this is good and that I should simplify as much as possible.

Let me take a break here for now. I do not think I will program today. I've mostly been deep in thought, and I'll let the decision settle in before tapping into it.

The reason why it takes me so long to decide is because I am bound by my old experiences. I've been tied to that in the ML library, and I am agitating over it here. But my ability is not the same as 1.5 years ago. It will take time to shed the rusty chains off me.