"5:40pm. I am back. I had a really good time to think it over. Is is possible to send a message and then block until the reply comes.
6:25pm. I am still thinking about it like crazy.

6:30pm. Still thinking about it.

Yeah, GUIs are a special case. I'll only ever need message passing for things like this between Spiral and F#. So using text for messaging should be fine. If I need to take screenshots at some point, I'll just do it in Spiral. GUIs are a huge rat's nets of OO that I will need the IDE's help to deal with.

Also, I will need to build up the `MailboxProcessor` and the `Async` modules, but that will be easy.

Let me stop here. I am not going to come to any deeper conclusions than this even if I continue thinking.

There is one thing that I realized though. I could have avoided putting the main loop of the previous poker game in async. Instead it really was possible to do it using continuations. Drat.

Well, live an learn. I did not have today's experience back then.

Actually, I do feel like grinding away at it for a while longer, but this does take a load of my shoulder. I do not have to bust my head on how I will do a GUI in Spiral anymore. I simply won't do it in it. With this crucial piece decided on, I can move on from here to getting back to work on the game.

6:55pm. Let me call it a day here. I really only planned like crazy, but that is fine. I'll get a bit dirty tomorrow."