I am really not sure what to think about that souped up architectures. To me it looks like it has 3 LSTMs all doing the same thing.
What the hell, I do not get it.

This is endemic of the RL field at this point it time.

Well, this is how I spent my morning. Let me get food and then after that I'll do some work.

Today my goal will be to make the game work using textual output and random players.

This is much like I've started the old thing in F#. Damn it, I really wish I did not have to leave that behind.

Spiral is trully the manifestation of my soul. It is greatly powerful, but at the same time useless and cannot play along well with others.

In 2-3 years, I'll have an awe inspiring power thanks to it, but for now nothing has changed. I need to pass through the desert first.