TODO: - Move the button on `next_round`. This will be done by rotating the lineup. - Make the internal representation function. - Make the deck. This one I can just port straightforwardly from the F# version.
Admittedly, I could have done all of this today had I put in more effort, but there is the GUI related despair to work through. I'll do it with with my usual effort and not a moment sooner.

I have plenty of things to be sad about. When all is said and done and the miracle happens, I will only have mastered only a tiny chunk of programming and will have left a mountain of work for my agents. That is the goal, but also the source of conflict.

How annoying. I went to so much trouble to make the GUI in the last version, but the thing is, I do not at all get the feeling that I need it anymore. It was so pretty, but it was also such a weight. A lot of complexity of the previous system was in that thing.

I want the GUI, but I also want to abstract all the effort in making it. I certainly do not want to spend months in giving Spiral the first rate GUI support. It was not at all what I had bargained for.

This inadequacy will continue to haunt me until I either do this, or I receive firm proof that GUIs are unneeded.

I will stop here for the day.