This should suffice to start the deck, but now it is time to make it properly.
Damn I'd rather read Goblin Slayer, but let me finish this and then I will take a break.

After I am done with this, all I would need are some random players and the output for which text will serve to the time being and I can run the game for the first time.

No doubt all this code will need debugging as well, but I will be done with it in a few days.

...Agh, let me just do it. This is pretty tedious.