I will chill a little more before I go on to debugging. I just do not feel like it.
Lately, I've been thinking about my past and about the stock market. It is really a major part of my life. If I play it smart and take the high road, I will never touch it again. If I do, in a fit of karmic retribution, I will surely lose everything.

One situation that is just too tempting to me that I need to avoid at all costs is betting on a market recovery after a crash.

The finanical markets are really nemesis of all intelligent things. It is a game designed to crush all individuality away. It does not even do it with rationality, but through irrationality. And the little pockets of individuality or rather, the opportunities where individuality could thrive are a few and far inbetween.

If I really want to trade, I should just make something and sell it. Zero risk of loss and a guaranteed return.

And my individuality will be made manifest in the final product.

That is the high road.

Poker is fine too. In it, the product is the gambling skill. The return is guaranteed to be riskless if I can pull off the act of creating the product.

The financial markets have their counters as well. It would not be tape reading or quick reflexes, but perfect understanding of conditions.

But the thing is - if I could create an agent capable of that, I surely woudn't waste my time using it to make financial bets. I'd just cause the Singularity and get an infinitely higher return than I could by speculating. I'd become the Future.

Ah damn it. It is like the markets are where dreams go to die. I will give that it is not even their fault this time. I should have had more foresight and trusted my own creativity and the power of machines.

This is the mistake I vow to not make again. Back then I have the excuse that I did not know better.

Now I do.

So the grand quest continues. Today I will finally move into testing the poker game.

For the time being let me chill a bit. I want to finish that Wiseman Pupil manga that I was reading yesterday and then I will start.