"Today is really full of distractions and I feel another essay coming on.
Let me make the Spiral resolution here. Now that the first part of the quest is complete, I am faced with challenges. Fundamentally, I am going to have to give up on some of the things I thought I would do.

I always take a very breadth first approach to things where I try to scoop up as much use as possible. But before long that approach will run into limits. In fact it is not humanly possible to take in everything.

Spiral is a special language, but it can't possibly be a good fit for everything. Here is its exact limit.

Imagine having to make a single page application that allow me to select models and set their parameters arbitrarily. Plus add multiple games to that for training models on them.

...It would be impossible to make such an application in Spiral. Not absolutely so, but it would never compile in sensible time. Spiral cannot be invoked in a JIT manner. It fundamentally needs to walk through the whole program. Hence it is unsuitable for making large programs without significant changes that I have no intention of implementing.

Fundamentally, Spiral will be limited to scripts.

It seems stupid to even think of it, but if I manage to do it, it will be done through extremely focused, highly tuned scripts. I will never get the polish I dreamed off with GUIs from Spiral. There would be too much to compile. Such flexiblity I will be only able to get by programming in the language directly.

The is the absolute limitation. Spiral does not have the explorability of a first rate statically typed language.

Spiral was pretty much created to cause the Singularity. It would be good for nothing else. It has unrivaled features for doing efficient parallel programing, writing ML libraries and games, but that is it.

I guess a language made to play with others would be a different beast from it after all. As much as it saddens me to admit, Spiral is not perfect.

It has a certain way of being used that is idiomatic to it.

It might seem like an absurd lie to say this at this point, but I've had absolute confidence that I'll be able to succeed even if these restrictions are in place. For one, GUIs are just vanity rather than an actual need. I sort of want to stack as much advantage as possible because I am that sort of person, but I can make do without it once I make up my mind.

So let me make this a resolution.

I'll keep running from here. I'll grasp the language that I had created and the methods that I forged tightly to me. And with it I will cause the Singularity in this time and in this age. Because if the Singularity cannot be cause; if Spiral's power is not enough to grasp it then it will most likely be impossible for anybody else as well and the world will enter a new dark age. Modernity will break upon its own weight. Games will either become power or they will perish along with their hosts.

I will not do it for others, but to save myself. The power of self improvement is only for the self - it cannot touch other people.

I have so many regrets now that I am almost 31 years of age, but my rage and my hatred will only be fuel for me doing the right thing - for myself. It is my way.

It is almost miraculous that Spiral would turn out like this. Though I would like it to be more, it exemplifies the selfish nature of the quest and the solitary path that lies ahead. Its true form is that of a battle I've long sought.

The above essay is a reflection of what was truly going on in the subconscious parts of my mind. I am about to partake in a great, enormous gamble and say that the Singularity will be caused by the exact, particular form of programming I am currently imagining and that nothing else is even remotely realistic.

Is it any wonder that I am scared and making only tepid steps towards the completion of the poker game. And that I am trying to delay by making plausible rationalizations for taking on side work like for the GUIs?

It is scary - if this fails I am going to face absolute denial of my existence. And if it succeeds I am not sure whether I will like what I am going to win.

'With this I have everything I need. Time to bring this to a finish.'

Underneath all the doubt and reason that tries to draw me towards safety is this one singular, terrifying thought that is as solid as a mountain and has all the weight of my soul in it.

'I cannot be stopped. There does not exist a reason to stop me.'

'With a few simple programs, I will end this world. They are not going to be very impressive programs, but they sure will be powerful.'

It is time to find them."

The above is the essay from the journal. Today these template troubles wrecked me.

I'll resume programming when I start feeling like it. It won't be long until I can honestly believe and then I will be able to draw out my full power. Right now, emotions are washing over me.