Damn it is still bright outside when though it is this late. I haven't been slacking at all.
I've been thinking. Basically things are going badly for me here. I remember a lot of high level details about how to do tabular RL, but I completely forgot how exactly it was implemented.

So when I started this two hours ago, I entered a deadlock regarding how I should proceed.

Yeah, it is true that I handled it last time, but I hadn't realized how poorly the whole thing had been done until now.

Crap. In isolation, this mutable way of doing the poker game might be fine, but having to do RL in addition to that is enough to push the whole thing into 'too hard to reason about' territory. Fucking `deck`. If that thing was not mutable I would not have gone with the current architecture. Really, just because I wanted to avoid allocating a single object, I let it dictate the course of the whole program. What a blunder.

And for a while there I started thinking I'd been getting better at this.

Well, redoing all of this will only take me a day or two at most so it is no big deal.

Let me just say this - having to do add RL to a game is the same as taking the gradients through the entire program. It is not a piece of cake. It needs to be a lot more flexible. I knew that and yet I ignored it.

Well, this is a usual error for me as I retrace my old steps. It happened with the ML library and it happened with the poker game.

For the time being, since I do not feel like starting the rewrite today, let me at least take care of the dictionary since I will need it to store the Q values.