Let me call it a day with this. I had enough for now.
Today really is not going the way I imagined it, but programming is more than just typing out code. It is imagination converted into power.

Make no mistake, doing the td-learner in a generic fashion for NL games would be a remarkable achievement.

I could not do it back in 2016. I might have done the first step of it, but not the second and the third, which is why the things that I wrote belong in the trash bin.

Yeah, I got some practive between early 2015-late 2016, but I wrote literally nothing of use during that period. It was like I set 1.5 years of my life on fire.

But of all the time I have wasted, the period of the last 3 years will turn out the absolutely best investment of my life.

Because now I have the ability to do it properly. And once that is done I will never have to touch the code again. It will be solid as a mountain.

To put it another way, sure, I will spend like extra two days fixing my mistake, but after that I will have something in my posession that would have then the old me a year and a half of work.

What I did in the last week cannot possibly be my true power.

That is coming right up when I am rested tomorrow.