5pm. Damn, it is this late. I've been in bed thinking.
I think I have most of the pieces now and in theory I could start work on it immediately. Actually, the amount of work I need to do is really small. RL is really a straightforward extension of supervised learning. There isn't that much difference.

But if I look at the actual field, the story is just so incoherent which sort of drives my doubts both towards the field and myself. So I am going to do just a tad more studying. I want to watch the videos by Levine again. DS's lectures are not actually that great on review.

Before I pick up the blade, I want to spend some time studying various RL algorithms.

The one thing that sticks to my mind is the experiment I did today where I tested square and binary CE again on the char RNN. I have absolutely no doubt that the reason softmax CE works while those don't is because it propagates errors through just one item. To be honest I'd have thought that the other two would work better that softmax CE, but that turns out not to be the case. Both softmax CE, dense connections, and dilated and temporal connections represent an important pieces of the sparsity puzzle. I feel that I really gained a lot by completing the char RNN.

A while ago I had an argument online where I said that rewards should be made discrete and one shot, but that can't be right.

Ilya is right - such a net would have very poor generalization...but...

No, he is right. So was that Reddit guy in the direction he was pushing in. No matter how I look at it, the reason the net blew up on the N puzzle is because of huge gradients at some step. I mean at the beginning of training it would not be strange for it to encounter target of 70 and the output of 3 and trying to backprop 67 caused it to blow up. Had I thought more clearly, I could have modified the cost function to clip the gradient to one.

Without a doubt, that would be the true solution. And actually isn't that the trick from TRPO or PPO? I am going to check this out. I can't keep clutching at one datapoint which was the N puzzle experiment without considering other evidence into consideration. I'll let poker decide it. If I manage to get scalar rewards in the output to work then probably there was a way to make them work with the N puzzle, only I missed it.

Also, apart from that, I did manage to solve the issue of how to deal with the replay buffer for sequences.

The way I am going to do it is the same as for char RNN? I am going to fold all the sequences into one and the minibatch them without worrying where one game ends and the other starts. No other approach apart from this is even remotely realistic. It is the only solution.

Really, there isn't that much work to making the RL player. I am agonizing right now about design choices, but I've already reasoned a lot of them out and that only leaves a tiny design space to search in compared to when I started. Understanding the gradient flow really helps a lot here. I'll get a significant amount of work done by the next update. It would not surprise me if in 2 months I am already thinking of how to raid online gambling lairs.

Enough dallying. Let me watch more videos and read some papers and then I will finish this. The second part of the grand quest will take nothing compared to the first.

It certainly won't take 3 years.