Dead tired from watching videos. Ok, now I am satisfied that I could implement both PG and Q learning with my eyes closed. I also got the convergence stuff too.
I need to get all of this thoroughly before I start.

I'll get it tomorrow. This is the final review before the decisive move and my focus is at its peak. Learning things in this state is a cinch.

It will be difficult to get this to work mostly because the algorithms are kind of shitty, but I will make it work somehow.

If nothing else, the thing most in my favor here is that poker is a reward rich game. It is much easier than real life. There is really no reason for me to fail at mastering it using reinforcement learning.

I'll go through all the necessary steps - one by one.