The experiment is a failure. But it does give me some insight into the nature of the problem which is prediction.
Policy gradients are prediction. Given how necessary the softmax is to the whole scheme and how much sense it makes, I'd say that it is a done deal.

But the way I am giong to do it right now does not make sense.

It only make sense to update the policy at the point of the reward. What I am going to essentially do is the wrong thing which is to sum it up and to hammer it at every step.

Honestly, it is no wonder that will make it take orders of magnitude more to converge.

This experiment just now was also enough to convince me that replacing the policy gradient update with regression on rewards has no chance of working.

It is really an obvious idea and if it had any chance of working, then somebody would have already tried it.

But the way to make RNNs work is non-obvious, so that is where I should focus my effort. It is not about statistics, but about reasoning out the correct architecture based on all the available evidence.

Since I am done with the videos and this, let me move on to reading the papers that I linked here earlier.