"6:25pm. I am calling it a day here.
This is enough of studies. I really want to put quote marks around "studies". It is not the problem with me, policy gradients and Q learning are just so uninspiring at this point.

I've am imagining there are three RL tiers above the current one before the Singularity. All the really fun stuff is in the last two.

This current one is just torture. The nets cannot reason, they cannot explore and they cannot generalize. In reverse order those 3 points correspond to tiers. In terms of generalization, I'd at least like to get the dataset munging out of the way. With supervised learning one can just do one hot encoding, but here there is only garbage.

I am not exactly sure what tier one will look like, but as a rough guide, evolutionary methods should be dead and buried at that stage. Random search beating deep RL is not exactly an inspirational story. Probably recurrent nets will come into their own in this stage. RL methods will become reliable and will no longer be reliant on hacks.

In tier two, being able to explore in a principled manner means there is no longer need to do repeated runs and as a result the sample efficiency would go way up. As a rough guide there should be a 10-100x gap in sample efficiency between each of the tiers. At this tier, programming ability becomes a weapon.

The third tier is the Ascent To Transcendence. It is the point where one is dealing with sentient agents before they become superhuman and the game changes completely because of that. In this stage my true talent will bloom, and it is where Homo Inspiratus will become separate from Homo Sapiens.

Probably the most defining feature of this tier is that agents will become capable of self-directed programming.

...

My head is full of noise. Understanding gradient propagation is just the tip of the iceberg. There is just so much more to this.

I won't hesitate any more. There is simply nothing more to be found in these papers. If there was then the field would highlight it. Policy gradient is my best bet. And the reality is that my ideas are absolutely juvenile. I am just full of shit. I should just get rid of my preconceptions and get the vanilla policy gradients to work. The rest of the methods are just hacks. And then once I do that on the one card poker game, I will progressively scale it, add RNNs to the mix and such. If possible, I'll somehow claw my way into tier one with my own power.

That is the way to do it.

Tomorrow I will follow the usual process when I plunge into the darkness - write out a sketch of the program and then once all the pieces are there, make the imagination real. That is my most powerful weapon as a programmer.

Despite the shittiness of the current methods, I should be able to make HU play work somehow because the episodes are short and there is no need to do much exploration due to the stochastic nature of the game. Literally everything is in my favor here, so I just need to man up and do it.

It will be done. This current stage is where I am absolutely most likely to face total failure. If I am commanding mastery of 6-man games at some point in the future, at that point I'll already know that I am on the right track.

Here there is a risk that I can crash and burn, and end up going into a programming job. I am sure there will be some employers impressed by my work on Spiral so it will serve as good resume item. I want to avoid that scenario at all cost. I'll give it a year to get something going. I've done everything to stack the deck in my favor and leveled up my programming abilities as far as I could. It might be a worthless talent, but it is still world class.

Even if I cannot believe in anything else, I will believe in it.

Tomorrow I will show my full power."