10:10am. https://arxiv.org/pdf/1802.04394.pdf ReinforceWalk- Learning to Walk in Graph with Monte Carlo Tree Search
Let me read this paper. I will absolutely do some programming today.

But first let me gather my nerve. Given that most of the users on the RL sub ignored the actually salient parts of my post to answer the trivial parts, I am starting to think they simply do not know. I have proded it again just now just to be sure, but it is unlikely that I will get the answer in the direction I am seeking, so I am going to have to go and get it myself.

In that case I think I will ignore both PG and Q-learning in favor of my own idea. I won't even try PG. I'll go with what I think is right from the start.

Also I wanted to say, that the experiment from yesterday was no a failure. With sigmoid activations in the decoder and sparse square, it does actually drive the error very low during training. It is just that unlike with softmax, what I get in the end is useless to sample from since only certain nodes are getting pushed around.

I mean if I have 0.95,0.95,1 in the decoder and the label is 0,0,1 that would get the perfect score under sparse squared error. Under softmax though this would also get the same score, but the '0.95's would get pushed towards zero. This is actually the behavior that is needed in order to get good sampling behavior. The sparse squared error is in fact optimizing as I thought it would so my hypothesis was in fact correct that sparsity is the reason it works at all.

https://arxiv.org/abs/1602.02867
Value Iteration Networks

Since what I am proposing to do is essentially value iteration with MC updates, let me read the above paper as well.