I think I will take a week off just to do studies of model based RL. I did not pay any attention to this part of the field at all and I need to fix that omission.
I will give up on my plan to make use of either PG or Q-learning. Originally I had the intention of using Q learning, but that was before I really thought things through.

For the past 3 years all my projects have essentially been hell waits on a remote possibility and this next part will not be any exception. The project will have many disparate pieces that I will have to assemble into one coherent whole. But the proof of concept of everything is already present in model based RL. I just need to make it work well and this is a great relief to me - that I can in fact get the sample complexity that I seek.

I will not seek to attain anything in the next month or in the next two months, but in 6-12 month I will have an absolute unbeatable player and a mastery of model based RL. I know how to do it, but the plan will need a lot of work to be brought to fruition.

The `poker5` test when I make it will not be a RL player, but a agent that learns the state dynamics. This is in fact what was gearing up to do 1.5 years ago. I should follow through with my plans instead of letting myself be swayed off course.

I'll continue watching the lectures and then read some Deepmind papers on control. Also I forgot to read the differentiable plasticity paper, so I should do that as well.

I need to better understand the strengths and weaknesses of the current state of the art and draw upon all the ideas before I go forward. Time to continue.