"4:30pm. Actually, I did realize something just recently. I do not need exploration in poker. It should arise naturally through the adversity of self play. If one of the players is too predicable the other will exploit it even with greedy action selection. There are no real local minima to hide in. Not in the HU games at least.
Actually, I had not realized this until now.

This makes a huge difference to everything.

...Actually, I think I know how to do it for multiplayer games as well through the hypernetworks approach. I see it now.

Rather than trying to bake exploration in, I should try to bake adversity in.

Yeah, that is it. With this is simple insight I essentially have what I need to solve the exploration exploitation dilemma for adversarial games - the most important kinds of games.

I do not really have to worry about tier two in practice.

I was actually worried about this. I knew that exploration was a large problem and that maybe the reason the value approach is not used is because it would lead to overfitting. I knew that the policy approach would maximize entropy and hence exploration. But here I do not have to do that at all!

I was actually extremely worried about this. Having this factor on hand actually makes it viable to do what I wanted to do. I was really beside myself.

Well, it was only in the past week that I started to seriously think of how I am going to do this."

This was 45m ago. I am really much more optimistic now. I have much more reason to think I will suceed at what I want to do.

I can think of so many application of adversity. Really it was such a stupid mistake for me to not realize this important thing even though it has been staring me in the face all along.