I was going to program today, but some interesting reading came up so I need to continue my studies. It is just as well. I'll start when I am ready.
http://etheses.whiterose.ac.uk/1315/
Rules of engagement- competitive coevolutionary dynamics in computational systems

This one is 210 pages, a PhD thesis.

https://pdfs.semanticscholar.org/18a0/38a670b48581555ed815774d84fad89c40dd.pdf
Coevolutionary Learning- a Case Study

https://www.reddit.com/r/reinforcementlearning/comments/8c94zb/recent_advancers_and_frontiers_in_deep_rl_mnih/?st=jfzq0m4h&sh=60b2c7b6

Ah, this is the Q-learning with the softmax that I've been saying is the right way of doing it this whole time.

It was called distributional RL.

https://pdfs.semanticscholar.org/9979/ababa4100cf35afc1c8be8777326134d14fd.pdf
Challenges in Coevolutionary Learning- Arms-Race Dynamics, Open-Endedness, and Mediocre Stable States