This is it! `Distributional Reinforcement Learning with Quantile Regression` is what I want to be using for updating the reward predictions. I am in absolute awe.
I had a lot of worries about the right way to update the net with reward labels, but these guys solved absolutely everything. I'll use this in the recurrent net and if that fails I'll use DQN with this.

I gained a truly great thing today.

Now, quantile regression is new to me, so I will have to spend some time studying it, including figuring out how to backprop through it, but that will be time well spent.

Let me move on to the following papers. I think NEC is next in line.