9am. I am up. I've done a lot of thinking as usual. Rather than give up, there is an option that can be used now that DRL with QR exists. Pseudocounts, intrinsic motivation, uncertainty estimates - those are all hacks to get around the fact that NNs cannot be optimistically initialized. The issue of exploration can be reduced to the issue of initialization.
What I really want is for the network to be optimistic on states it has not visited before and less so on the ones it has just like in the tabular case.

Things like jacking the bias in the top layer would not work because it would be quickly undoed by gradient descent. What really should be done is figure out a scheme to do it in the intermediate layers. This would ordinarily be hard, but there are some clues that lead not necessarily how to do it, but how not to do it.

A roboust method to do it would not jack the biases. Layer normalization should be used for stability and it would just recenter it and that point of doing that would be lost. So necessarily, a roboust method would operate on the weights. But is it really possible to make a roboust initialization method that is optimistic on all possible states before it has been seen any of them?

Honestly, I have no idea how to do that.

So if we give up on messing with the weights and messing with the biases that only leaves one thing - shifting the reward predictions by some large positive constant and forcing the network to adapt to it. This is not the same as shifting the reward itself.

Now...before concluding whether that is right or not, I had an idea that maybe most of the learning in RL domains is in fact noise learning.

https://arxiv.org/abs/1704.05310
Unsupervised Learning by Predicting Noise

https://arxiv.org/abs/1706.05394
A Closer Look at Memorization in Deep Networks

Let me read these two papers which I got referred to based on my question on the ML sub yesterday.