"3:20pm. Focus, just relax for a bit.
Unwind. This is the last thing.

I just need to pay attention and the way to do it all will fall into my lap.

3:35pm. No, I can't focus on Nanoha.

Let me do some more thinking. Yeah, all the pieces are there almost. the `Q(s,a) -> p(a)` function is missing though.

3:40pm. http://artint.info/html/ArtInt_266.html

Yeah, the softmax would be the most reasonable choice as usual here.

Yeah, I will just go with that. No need to ditch everything that I have and go with PG just because I do not want to set the hyperparameters of this little thing.

There is an interesting thing here though. As T goes to zero so the gradient gate would get shut.

Actually rather than making T a hyperparameter I could always optimize it using backprop as well. I need to learn from the differentiable plasticity thing. Why not give the net a chance to control its own exploration?

I do not have to be restricted in any way whatsoever. I can multiply the Q(s,a) outputs by a matrix before passing it into the softmax. The action selection can be done after calculating the Q values. Only during the backward phase would knowing which action to propagate back through would be important. And only in the final step would Q values need to be optimized directly using the cost function.

Actually if I do it like this, then at the start of training I won't need to do that optimistic init hack. The Boltzman action selection should do the trick.

3:55pm. This will work. It has to work. The way to do it to let training be as imaginary as possible.

I will not actually let the net see the true state except at the beginning. In the following steps I will integrate the imagined state and the true state using attention like mechanism. It will just be multiplicative integration. It feels like a smarter idea than just forcing the net to predict the next state. Doing it integratively will force the net to ...

...Actually, doing it like that would not be world modeling. It would just be attention. ...Which is actually better, if I can make it work.

This will connect everything into one whole and will allow me to reach the threshold to the first tier. It will be the first spark of the inspired desire.

Probably that guy who said that the optimistic init will get corrected is right. But what I am doing here is a very specific thing that needs one to focus directly on poker in order to get it to work. It is not that it has to be poker, but poker is the ideal environment to cook this up.

4:05pm. Now can I rest?

...No. I want to study attention a little."