I'll stop here for a bit so I can eat.
I am getting good feedback as a I program. It was not like a week ago when I tried to force it even though I was full of doubt.

I can do this.

Ultimately, there are not that many paths I can take to making a good player using RL. It is just a matter of scaling at this point. And as time goes on, the whole field will get better taking me along for the right.