```     inl RL =         inl dq_net =             inl net =                 open Feedforward.Layer                 inl label = input .label hidden_size                 inl network =                     input .input input_size                     |> linear hidden_size                     |> init s                 error Error.square label network
        {dq_net}
```

I need to make special input layers for this.

There is one thing that is coming to mind for me though. Rather than needing modulation, if I consider the hidden state as a part of state, maybe the RNN would be perfectly capable of modulating on its own even if MC returns were given at every timestep.

I am thinking of the case where no input is given at all and the network needs to fly blind. This would be the pure reasoning scenario.

It does not really make sense that a RNN would need some kind of hack like a modulator.

Yeah, it should be as simple as possible and no simpler. I'll know when I get some practical experience. I do not expect I will get the ideal scenario with and out of the box RNN. It will require some kind of architectural improvement.

Let me work on the input layers. This will be the big and major thing that I've made Spiral for. Being able to reflect on the input type and one-hot encode it was one of the pain points last time. It will save me a lot of bother.